Move display function to display class
