+++ tests/config.yaml
@@ -5,8 +5,10 @@ enc:
   origin: [ 17.5, 54.0 ]
   depths: [ 0, 1, 2, 5, 10, 20, 50, 100, 200, 350, 500 ]
   crs: "WGS84"
-  S57_layers: ["DEPARE", "LNDARE"]
-  resources: [ "data/", "data/db/PL2MP500/4/0" ]
+  S57_layers:
+    "TSSLPT": "#8B0000"
+    "CTNARE": "#964B00"
+  resources: [ "data/db/PL2MP500/4/0" ]
 #  weather:
 #    PyThor_adress: "http://127.0.0.1:5000"
 #    time_start: 1717931596
@@ -19,11 +21,11 @@ enc:
 #  depths: [ 0, 1, 2, 5, 10, 20, 50, 100, 200, 350, 500 ]
 #  crs: "UTM33N"
 #  resources: [ "/", "data/", "data/db/" ]
-  # ais:
-  #   module: "live"
-  #   address: "195.182.206.250"
-  #   port: 54321
-  #   interval: 10
+  ais:
+    module: "live"
+    address: "195.182.206.250"
+    port: 54321
+    interval: 10
 display:
   colorbar: False
   dark_mode: True
