+++ tests/test_s57.py
@@ -4,11 +4,12 @@ from seacharts.core.extent import Extent
 from pyproj import Transformer
 
 enc = ENC("config.yaml")
+# ships = [
+#         (1, 612291, 6071542, 132, "orange"),
+#         (2, 356393, 6052365, 57, "yellow")
+#     ]
 
-print(enc.seabed[10])
-print(enc.shore)
-print(enc.land)
-# ship = (1,612291 ,6071542,96,"red")
-# ships = [tuple(ship), ]
 # enc.display.add_vessels(*ships)
+# enc.display.start()
+
 enc.display.show()
